Split smart message handler for different FHIR versions
