+++ README.md
@@ -8,40 +8,54 @@ A Java port of the [SMARTWebEHR](https://github.com/Tiro-health/SMARTWebEHR) .NE
 - SDC (Structured Data Capture) operations
 - Form submission handling
 - Event-driven architecture
-- HAPI FHIR R5 integration
+- Multi-version FHIR support (R4 and R5)
 
 ## Installation
 
-Available on [Maven Central](https://repo1.maven.org/maven2/health/tiro/smart-web-ehr/).
+Available on [Maven Central](https://repo1.maven.org/maven2/health/tiro/).
 
-Add to your `pom.xml`:
+Pick the module matching your FHIR version:
+
+**FHIR R5:**
 
 ```xml
 <dependency>
     <groupId>health.tiro</groupId>
-    <artifactId>smart-web-ehr</artifactId>
-    <version>1.0.7</version>
+    <artifactId>smart-web-ehr-r5</artifactId>
+    <version>2.0.0</version>
 </dependency>
 ```
+
+**FHIR R4:**
+
+```xml
+<dependency>
+    <groupId>health.tiro</groupId>
+    <artifactId>smart-web-ehr-r4</artifactId>
+    <version>2.0.0</version>
+</dependency>
+```
+
+Both modules transitively include `smart-web-ehr-core`, so you only need one dependency.
 
 ## Usage
 
 ### Basic Setup
 
 ```java
-import health.tiro.smartwebehr.SmartMessageHandler;
-import health.tiro.smartwebehr.events.*;
+// R5 example — for R4, change the import to health.tiro.smartwebehr.r4.SmartMessageHandler
+import health.tiro.smartwebehr.r5.SmartMessageHandler;
+import health.tiro.smartwebehr.r5.R5SmartMessageListener;
+import org.hl7.fhir.r5.model.*;
 
 // Create handler
 SmartMessageHandler handler = new SmartMessageHandler();
 
-// Add event listeners
-handler.addListener(new SmartMessageListener() {
+// Add event listeners (typed adapter — no casts needed)
+handler.addListener(new R5SmartMessageListener() {
     @Override
-    public void onFormSubmitted(FormSubmittedEvent event) {
-        QuestionnaireResponse response = event.getResponse();
-        OperationOutcome outcome = event.getOutcome();
-        // Process the submitted form
+    public void onFormSubmitted(QuestionnaireResponse response, OperationOutcome outcome) {
+        // Process the submitted form — fully typed
     }
 
     @Override
@@ -59,13 +73,13 @@ import com.teamdev.jxbrowser.engine.Engine;
 import com.teamdev.jxbrowser.js.JsObject;
 
 public class JxBrowserIntegration {
-    
+
     private SmartMessageHandler handler;
     private Browser browser;
-    
+
     public void setup() {
         handler = new SmartMessageHandler();
-        
+
         // Set up message sender for outbound messages
         handler.setMessageSender(jsonMessage -> {
             CompletableFuture<String> future = new CompletableFuture<>();
@@ -75,7 +89,7 @@ public class JxBrowserIntegration {
             });
             return future;
         });
-        
+
         // Inject message handler into browser
         browser.mainFrame().ifPresent(frame -> {
             JsObject window = frame.executeJavaScript("window");
@@ -86,7 +100,7 @@ public class JxBrowserIntegration {
                 }
             });
         });
-        
+
         // Add listener for incoming messages
         browser.mainFrame().ifPresent(frame -> {
             frame.executeJavaScript(
@@ -107,13 +121,13 @@ public class JxBrowserIntegration {
 import com.nicepay.nicefx.browser.ChromiumBrowser;
 
 public class EquoIntegration {
-    
+
     private SmartMessageHandler handler;
     private ChromiumBrowser browser;
-    
+
     public void setup() {
         handler = new SmartMessageHandler();
-        
+
         // Set up message sender
         handler.setMessageSender(jsonMessage -> {
             CompletableFuture<String> future = new CompletableFuture<>();
@@ -121,7 +135,7 @@ public class EquoIntegration {
             future.complete(null);
             return future;
         });
-        
+
         // Register Java function to be called from JavaScript
         browser.registerBrowserFunction("smartMessageCallback", args -> {
             if (args.length > 0) {
@@ -133,7 +147,7 @@ public class EquoIntegration {
             }
             return null;
         });
-        
+
         // Set up message listener in browser
         browser.executeJavaScript(
             "window.addEventListener('message', function(event) {" +
@@ -181,6 +195,14 @@ handler.sendSdcConfigureContextAsync(
 );
 ```
 
+## Module Structure
+
+| Module | Artifact | Description |
+|--------|----------|-------------|
+| Core | `smart-web-ehr-core` | Shared logic, FHIR-version-independent. Depends on `hapi-fhir-base` only. |
+| R4 | `smart-web-ehr-r4` | FHIR R4 typed API. Depends on core + `hapi-fhir-structures-r4`. |
+| R5 | `smart-web-ehr-r5` | FHIR R5 typed API. Depends on core + `hapi-fhir-structures-r5`. |
+
 ## Message Types Supported
 
 ### Inbound (from WebView)
@@ -195,13 +217,28 @@ handler.sendSdcConfigureContextAsync(
 - `sdc.configureContext` - Configure launch context
 - `sdc.displayQuestionnaire` - Display a questionnaire
 
+## Migrating from 1.x
+
+```java
+// Before (1.x)
+import health.tiro.smartwebehr.SmartMessageHandler;
+
+// After (2.x) — pick your FHIR version
+import health.tiro.smartwebehr.r5.SmartMessageHandler;  // R5
+import health.tiro.smartwebehr.r4.SmartMessageHandler;  // R4
+```
+
+The `SmartMessageHandler` API is unchanged. The only difference is the import and the Maven artifact.
+
+For typed event access, use `R5SmartMessageListener` or `R4SmartMessageListener` instead of `SmartMessageListener`.
+
 ## Requirements
 
 - Java 8 or higher
 
 ## Dependencies
 
-- HAPI FHIR R5 (5.7.0)
+- HAPI FHIR (5.7.0)
 - Jackson Databind (2.15.3)
 - SLF4J (2.0.9)
 - Javax Validation API (2.0.1.Final)
