+++ smart-web-ehr-core/src/main/java/health/tiro/smartwebehr/events/CloseApplicationEvent.java
