+++ smart-web-ehr-core/src/main/java/health/tiro/smartwebehr/message/SmartMessageBase.java
