+++ smart-web-ehr-core/src/main/java/health/tiro/smartwebehr/message/payload/ErrorResponse.java
