+++ smart-web-ehr-r4/pom.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>health.tiro</groupId>
+        <artifactId>smart-web-ehr-parent</artifactId>
+        <version>2.0.0</version>
+    </parent>
+
+    <artifactId>smart-web-ehr-r4</artifactId>
+    <packaging>jar</packaging>
+
+    <name>SMART Web EHR R4</name>
+    <description>SMART Web Messaging for HAPI FHIR R4</description>
+
+    <dependencies>
+        <dependency>
+            <groupId>health.tiro</groupId>
+            <artifactId>smart-web-ehr-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>ca.uhn.hapi.fhir</groupId>
+            <artifactId>hapi-fhir-structures-r4</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+</project>
